pyproject.toml
[tool.poetry]
name = "mimesis_factory"
version = "0.1.0"
description = "Mimesis integration with factory_boy"

license = "MIT"

authors = [
    "Nikita Sobolev <mail@sobolevn.me>",
    "Líkið Geimfari <likid.geimfari@gmail.com>"
]

readme = 'README.md'  # Markdown files are supported

repository = "https://github.com/mimesis-lab/mimesis-factory"
homepage = "https://github.com/mimesis-lab/mimesis-factory"

keywords = ["mimesis", "factory_boy", "testing"]

[tool.poetry.dependencies]
python = "^3.5 || ^3.6"
factory-boy = ">=2.10"
mimesis = ">=2.0"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.5"
pytest-isort = "^0.2.0"
pytest-flake8 = "^1.0"
pytest = "^3.6"
flake8-builtins = "^1.4"
flake8-commas = "^2.0"
flake8-quotes = "^1.0"
flake8-pep3101 = "^1.2"
flake8-comprehensions = "^1.4"
flake8-blind-except = "^0.1.1"
flake8-docstrings = "^1.3"
flake8-string-format = "^0.2.3"
flake8-coding = "^1.3"
flake8-module-name = "^0.1.5"
flake8-bugbear = "^18.2"
flake8-pytest = "^1.3"
pep8-naming = "^0.7.0"
flake8 = "^3.5"
validators = "^0.12.2"
pytest_factoryboy = "^2.0"
pytest-randomly = "^1.2"
